+++ packages/core/src/tools/burn-subtitles/index.ts
@@ -101,7 +101,7 @@ export const burnSubtitles: ToolModule<BurnSubtitlesParams> = {
     ]);
 
     const output = await ff.readFile(outputName);
-    const outputBytes: Uint8Array = typeof output === 'string' ? new TextEncoder().encode(output) : (output as Uint8Array);
+    const outputBytes: Uint8Array = typeof output === 'string' ? new TextEncoder().encode(output) : (output);
 
     await ff.deleteFile(inputName);
     await ff.deleteFile(subName);
