+++ packages/core/src/tools/convert-video/index.ts
@@ -132,7 +132,7 @@ export const convertVideo: ToolModule<ConvertVideoParams> = {
     await ff.deleteFile(outputName);
     const outputBytes = typeof output === 'string'
       ? new TextEncoder().encode(output)
-      : (output as Uint8Array);
+      : (output);
 
     ctx.onProgress({ stage: 'done', percent: 100, message: 'Done' });
     return [new Blob([outputBytes.buffer as ArrayBuffer], { type: getVideoMime(params.format) })];
