+++ packages/core/src/tools/letterbox-video/index.ts
@@ -101,7 +101,7 @@ export const letterboxVideo: ToolModule<LetterboxVideoParams> = {
     await ff.exec(buildLetterboxArgs(inputName, outputName, params.aspect));
     const output = await ff.readFile(outputName);
     const outputBytes: Uint8Array =
-      typeof output === 'string' ? new TextEncoder().encode(output) : (output as Uint8Array);
+      typeof output === 'string' ? new TextEncoder().encode(output) : (output);
 
     await ff.deleteFile(inputName).catch(() => {});
     await ff.deleteFile(outputName).catch(() => {});
