+++ packages/core/src/tools/video-side-by-side/index.ts
@@ -92,7 +92,7 @@ export const videoSideBySide: ToolModule<VideoSideBySideParams> = {
     await ff.exec(buildSideBySideArgs(aName, bName, outputName, params.orientation));
     const output = await ff.readFile(outputName);
     const outputBytes: Uint8Array =
-      typeof output === 'string' ? new TextEncoder().encode(output) : (output as Uint8Array);
+      typeof output === 'string' ? new TextEncoder().encode(output) : (output);
 
     await ff.deleteFile(aName).catch(() => {});
     await ff.deleteFile(bName).catch(() => {});
