+++ packages/core/src/tools/vignette-video/index.ts
@@ -88,7 +88,7 @@ export const vignetteVideo: ToolModule<VignetteVideoParams> = {
     await ff.exec(buildVignetteArgs(inputName, outputName, params.strength));
     const output = await ff.readFile(outputName);
     const outputBytes: Uint8Array =
-      typeof output === 'string' ? new TextEncoder().encode(output) : (output as Uint8Array);
+      typeof output === 'string' ? new TextEncoder().encode(output) : (output);
 
     await ff.deleteFile(inputName).catch(() => {});
     await ff.deleteFile(outputName).catch(() => {});
